Update base.py
RuntimeError: "host_softmax" not implemented for 'Int'

+++ lab3/text_recognizer/lit_models/base.py
@@ -21,7 +21,7 @@ class Accuracy(pl.metrics.Accuracy):
         Normalized preds are not necessary for accuracy computation as we just care about argmax().
         """
         if preds.min() < 0 or preds.max() > 1:
-            preds = torch.nn.functional.softmax(preds, dim=-1)
+            preds = torch.nn.functional.softmax(preds.float(), dim=-1) # RuntimeError: "host_softmax" not implemented for 'Int'
         super().update(preds=preds, target=target)
